39836 - Criada nova variável e ajustado modo de leitura base64
