+++ src/SME.SR.Infra/Utilitarios/VariaveisAmbiente.cs
@@ -11,5 +11,6 @@ namespace SME.SR.Infra
         public string ConnectionStringCoreSso => Environment.GetEnvironmentVariable("CoreSSOConnection");
         public string ConnectionStringSondagem => Environment.GetEnvironmentVariable("sondagemConnection");
         public string ConnectionStringAE => Environment.GetEnvironmentVariable("AEConnection");
+        public string PastaArquivosSGP => Environment.GetEnvironmentVariable("SGPPastaArquivos");
     }
 }
